Rough quad scanner, bug ends in "face-select" mode
Able to run a scan through all faces of geometry present in a scene and report how many triangles, quadrilaterals, and n_gons there are. Lingering bug however in that the execution always ends with the selection mode for all objects set to face mode instead of whatever it may have been prior to execution.